requirements.txt
# requirements.txt (Ensure these versions, especially cryptography)
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.12.3
bleak==0.22.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
email_validator>=1.1
filelock==3.18.0
FindMy==0.7.6
Flask==3.1.0
Flask-Executor==1.0.0
Flask-Login==0.6.3
Flask-WTF>=1.0.0
frozenlist==1.5.0
h11==0.14.0
http_ece==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
py-vapid==1.9.2
pycparser==2.22
python-dotenv==1.0.1
pywebpush==2.0.3
requests==2.32.3
six==1.17.0
soupsieve==2.6
srp==1.0.22
typing_extensions==4.12.2
urllib3==2.3.0
# uvicorn not strictly needed if using Waitress/Gunicorn
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.18.3
APScheduler>=3.9.0 # Ensure version
regex==2024.11.6 # optional for generate_svg, can be added if needed
# geopy removed, using manual haversine